update:Post like function

+++ backend/src/main/java/com/devu/backend/controller/post/ResponsePostDto.java
@@ -20,7 +20,7 @@ public class ResponsePostDto {
     private String content;
     private List<String> url;
     private Long hit;
-    private Long like;
+    private Integer like;
     private StudyStatus studyStatus;
     private QuestionStatus questionStatus;
 }
